Refactor jpa test packages.
