+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/jpa/DummyEntityRESTResource.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.test.clustering.cluster.jpa2lc;
+package org.jboss.as.test.clustering.cluster.jpa;
 
 import javax.ejb.LocalBean;
 import javax.ejb.Stateless;
