+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/jpa/custom/ClusteredJPA2LCCustomRegionTestCase.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.test.clustering.cluster.jpa2lcCustomRegion;
+package org.jboss.as.test.clustering.cluster.jpa.custom;
 
 import org.jboss.arquillian.container.test.api.ContainerController;
 import org.jboss.arquillian.container.test.api.Deployer;
