+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/jpa/custom/DummyEntityControllingApplication.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.test.clustering.cluster.jpa2lcCustomRegion;
+package org.jboss.as.test.clustering.cluster.jpa.custom;
 
 import javax.ws.rs.ApplicationPath;
 import java.util.HashSet;
