+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/jpa/persistence.xml
